IN-361 fix: fix payments pagination
