+++ src/features/payments/model/useMyPayments.ts
@@ -1,15 +1,26 @@
 import { useEffect, useState } from 'react'
 
+import { useAppSelector } from '@/common/lib/hooks/reduxHooks'
+import { authSlice } from '@/entities/auth'
 import { MyPayment, paymentsApi } from '@/entities/payments'
+import { TableData } from '@byte-creators/ui-kit'
+import { useScopedTranslation } from '@byte-creators/utils'
+import { useRouter } from 'next/router'
 
 export const useMyPayments = () => {
   const { data, isLoading } = paymentsApi.useGetMyPaymentsQuery()
+
+  const t = useScopedTranslation('Payments')
+  const router = useRouter()
+  const userId = useAppSelector(authSlice.selectors.selectUserId)
+  const step = 6
   const [currentPage, setCurrentPage] = useState(1)
-  const [dataPortion, setDataPortion] = useState(10)
+  const [dataPortion, setDataPortion] = useState(step)
   const [dataForDisplay, setDataForDisplay] = useState<MyPayment[] | undefined>([])
 
   const [paymentSuccess, setPaymentSuccess] = useState(false)
   const [paymentFailed, setPaymentFailed] = useState(false)
+  const pagesPortionOptions = ['6', '8', '10', '20', '30', '50']
 
   useEffect(() => {
     setDataForDisplay(data?.slice(0, dataPortion))
@@ -41,6 +52,20 @@ export const useMyPayments = () => {
     pagesCount = Math.ceil(data.length / dataPortion)
   }
 
+  let payments = [] as TableData[]
+
+  if (dataForDisplay) {
+    payments = dataForDisplay?.map((el: MyPayment) => {
+      return {
+        1: new Date(el.dateOfPayment).toLocaleDateString(),
+        2: new Date(el.endDateOfSubscription).toLocaleDateString(),
+        3: `$${el.price}`,
+        4: el.subscriptionType,
+        5: el.paymentType,
+      }
+    })
+  }
+
   return {
     currentPage,
     dataForDisplay,
@@ -49,9 +74,14 @@ export const useMyPayments = () => {
     handlePaymentsPortionChange,
     isLoading,
     pagesCount,
+    pagesPortionOptions,
     paymentFailed,
     paymentSuccess,
+    payments,
+    router,
     setPaymentFailed,
     setPaymentSuccess,
+    t,
+    userId,
   }
 }
