+++ src/features/payments/ui/MyPayments.tsx
@@ -1,14 +1,11 @@
 import React, { useEffect } from 'react'
 
-import { MyPayment } from '@/entities/payments'
-import { Loader, Pagination, Table, TableData, Typography } from '@byte-creators/ui-kit'
-import { useScopedTranslation } from '@byte-creators/utils'
-import { useRouter } from 'next/router'
+import { Button, Loader, Pagination, Table, Typography } from '@byte-creators/ui-kit'
+import { ArrowBackOutline } from '@byte-creators/ui-kit/icons'
 
 import { useMyPayments } from '../model/useMyPayments'
 
 export const MyPayments = () => {
-  const t = useScopedTranslation('Payments')
   const {
     currentPage,
     dataForDisplay,
@@ -17,13 +14,16 @@ export const MyPayments = () => {
     handlePaymentsPortionChange,
     isLoading,
     pagesCount,
+    pagesPortionOptions,
     paymentFailed,
     paymentSuccess,
+    payments,
+    router,
     setPaymentFailed,
     setPaymentSuccess,
+    t,
+    userId,
   } = useMyPayments()
-
-  const router = useRouter()
 
   useEffect(() => {
     if (router.query.success === 'true') {
@@ -34,20 +34,6 @@ export const MyPayments = () => {
       setPaymentFailed(true)
     }
   }, [router.query.success])
-
-  let payments = [] as TableData[]
-
-  if (dataForDisplay) {
-    payments = dataForDisplay?.map((el: MyPayment) => {
-      return {
-        1: new Date(el.dateOfPayment).toLocaleDateString(),
-        2: new Date(el.endDateOfSubscription).toLocaleDateString(),
-        3: `$${el.price}`,
-        4: el.subscriptionType,
-        5: el.paymentType,
-      }
-    })
-  }
 
   const headers = [
     {
@@ -73,18 +59,29 @@ export const MyPayments = () => {
     </div>
   )
 
-  const renderEmptyMessage = () => <Typography>You do not have any subscriptions yet</Typography>
+  const renderEmptyMessage = () => (
+    <div className={'flex justify-center items-center'}>
+      <Button
+        className={'bg-transparent'}
+        onClick={() => router.push(`/profile/${userId}/settings`)}
+        variant={'icon'}
+      >
+        <ArrowBackOutline />
+      </Button>
+      <Typography>You do not have any subscriptions yet</Typography>
+    </div>
+  )
 
   const renderTableWithPagination = () => (
     <div>
       <Table headers={headers} tableData={payments} />
       <Pagination
         className={'hidden sm:inline-flex sm:self-start sm:w-auto sm:mt-9 sm:mb-16'}
-        currentPage={1}
-        onChangePagesPortion={() => {}}
-        onClickPaginationButton={() => {}}
-        pagesCount={2}
-        pagesPortion={'5'}
+        currentPage={currentPage}
+        onChangePagesPortion={handlePaymentsPortionChange}
+        onClickPaginationButton={handleCurrentPageChange}
+        pagesCount={pagesCount}
+        pagesPortionOptions={pagesPortionOptions}
       />
     </div>
   )
